requirements.txt
pandas==2.2.0
regex==2023.12.25
openpyxl==3.1.2
emoji==2.10.1
torch==2.2.0
scikit-learn==1.4.0
transformers==4.37.2
py_vncorenlp==0.1.4
google-api-python-client==2.116.0
python-dotenv==1.0.1
boto3==1.34.34
pypyr==5.9.1
pyarrow==15.0.0
fasttext-langdetect==1.0.5
fastapi==0.109.2
uvicorn==0.27.0.post1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9